news-data/requirements.txt
feedparser
mysql-connector-python
python-dotenv
beautifulsoup4
requests
numpy
sentence-transformers
scikit-learn
tqdm
python-dateutil
pymysql
